# backend/openapi.json
{
  "components": {
    "schemas": {
      "Appointment": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "referral_id": {
            "title": "Referral Id",
            "type": "string"
          },
          "slot": {
            "$ref": "#/components/schemas/AppointmentSlot"
          },
          "status": {
            "$ref": "#/components/schemas/ReferralStatus"
          }
        },
        "required": [
          "id",
          "referral_id",
          "slot",
          "status"
        ],
        "title": "Appointment",
        "type": "object"
      },
...
      "ApproveRequest": {
        "description": "The physician's explicit approval. Booking happens only as a result of this call.",
        "properties": {
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "match_run_id": {
            "title": "Match Run Id",
            "type": "string"
          },
          "slot_id": {
            "title": "Slot Id",
            "type": "string"
          },
          "specialist_id": {
            "title": "Specialist Id",
            "type": "string"
          }
        },
        "required": [
          "match_run_id",
          "specialist_id",
          "slot_id"
        ],
        "title": "ApproveRequest",
        "type": "object"
      },
      "ApproveTrialRequest": {
        "description": "The physician's explicit approval to surface this trial to the patient.",
        "properties": {
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          }
        },
        "title": "ApproveTrialRequest",
        "type": "object"
      },
...
      "Complexity": {
        "description": "Claude proposes this at parse time; the physician confirms or overrides before matching.",
        "enum": [
          "routine",
          "rare_complex"
        ],
        "title": "Complexity",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchRun": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "referral_id": {
            "title": "Referral Id",
            "type": "string"
          },
          "results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MatchResult"
            },
            "title": "Results",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "too_late": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TooLateCandidate"
            },
            "title": "Too Late",
            "type": "array"
          },
          "weight_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeightProfile"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "referral_id",
          "status",
          "created_at"
        ],
        "title": "MatchRun",
        "type": "object"
      },
...
      "PatientSummary": {
        "properties": {
          "appointment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Appointment"
              },
              {
                "type": "null"
              }
            ]
          },
          "cost_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Note"
          },
          "referral_id": {
            "title": "Referral Id",
            "type": "string"
          },
          "specialist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialist Name"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "trial_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Note"
          }
        },
        "required": [
          "referral_id",
          "summary"
        ],
        "title": "PatientSummary",
        "type": "object"
      },
...
      "Referral": {
        "properties": {
          "appointment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Appointment"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Approval"
              },
              {
                "type": "null"
              }
            ]
          },
          "case_notes": {
            "title": "Case Notes",
            "type": "string"
          },
          "complexity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Complexity"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "parsed_case": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedCase"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "referring_physician_id": {
            "title": "Referring Physician Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReferralStatus"
          },
          "trial_approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrialApproval"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/Urgency"
          }
        },
        "required": [
          "id",
          "patient_id",
          "referring_physician_id",
          "case_notes",
          "urgency",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Referral",
        "type": "object"
      },
      "ReferralCreate": {
        "examples": [
          {
            "case_notes": "58F, 3 months progressive right knee pain, worse on stairs. Swelling after activity. No trauma. X-ray shows joint space narrowing.",
            "patient_id": "pat_001",
            "urgency": "routine"
          }
        ],
        "properties": {
          "case_notes": {
            "minLength": 1,
            "title": "Case Notes",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "referring_physician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referring Physician Id"
          },
          "urgency": {
            "$ref": "#/components/schemas/Urgency",
            "default": "routine"
          }
        },
        "required": [
          "patient_id",
          "case_notes"
        ],
        "title": "ReferralCreate",
        "type": "object"
      },
      "ReferralStatus": {
        "description": "approved/booked require an explicit physician action.",
        "enum": [
          "draft",
          "matched",
          "approved",
          "booked",
          "waitlisted",
          "cancelled"
        ],
        "title": "ReferralStatus",
        "type": "string"
      },
      "ReferralUpdate": {
        "description": "Physician confirmation/override. Matching requires both to be set (409 otherwise).",
        "properties": {
          "complexity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Complexity"
              },
              {
                "type": "null"
              }
            ]
          },
          "urgency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Urgency"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ReferralUpdate",
        "type": "object"
      },
...
      "Trial": {
        "properties": {
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "condition_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Condition Keywords",
            "type": "array"
          },
          "eligibility_summary": {
            "title": "Eligibility Summary",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "nct_id": {
            "title": "Nct Id",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nct_id",
          "title",
          "phase",
          "condition",
          "condition_keywords",
          "summary",
          "eligibility_summary",
          "location"
        ],
        "title": "Trial",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Dev mode only (AUTH_MODE=dev): act as this physician id, e.g. doc_001",
        "in": "header",
        "name": "X-Dev-Physician",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "description": "Auth0 access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Physician-first referral matching. Nothing is sent or booked without physician approval.",
    "title": "ECHO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/referrals": {
      "get": {
        "description": "The signed-in physician's referrals (an admin sees everyone's).",
        "operationId": "list_referrals_referrals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Referral"
                  },
                  "title": "Response List Referrals Referrals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Referrals",
        "tags": [
          "referrals"
        ]
      },
      "post": {
        "operationId": "create_referral_referrals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Referral",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}": {
      "get": {
        "operationId": "get_referral_referrals__referral_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Referral",
        "tags": [
          "referrals"
        ]
      },
      "patch": {
        "description": "Physician confirms or overrides urgency and complexity.",
        "operationId": "confirm_referral_referrals__referral_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Confirm Referral",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}/approve": {
      "post": {
        "description": "The physician's explicit approval. This is the only path that approves or books.",
        "operationId": "approve_and_book_referrals__referral_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Approver is not the referring physician"
          },
          "409": {
            "description": "Stale match run, or slot no longer available; re-run matching"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Approve And Book",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}/match": {
      "get": {
        "description": "Latest match run for the referral; poll until status is complete or failed.",
        "operationId": "get_latest_match_referrals__referral_id__match_get",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Latest Match",
        "tags": [
          "referrals"
        ]
      },
      "post": {
        "operationId": "start_match_referrals__referral_id__match_post",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Not parsed, or complexity not yet confirmed by the physician"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Match",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}/parse": {
      "post": {
        "description": "Queue case parsing. Poll GET /referrals/{id} until `parsed_case` is set.",
        "operationId": "parse_referral_referrals__referral_id__parse_post",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Parse Referral",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}/patient-summary": {
      "get": {
        "operationId": "patient_summary_referrals__referral_id__patient_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Patient Summary",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}/trials": {
      "get": {
        "operationId": "list_trials_referrals__referral_id__trials_get",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Trial"
                  },
                  "title": "Response List Trials Referrals  Referral Id  Trials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Not parsed, or complexity is not rare_complex"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Trials",
        "tags": [
          "trials"
        ]
      }
    },
    "/referrals/{referral_id}/trials/{trial_id}/approve": {
      "post": {
        "description": "The physician's explicit approval to discuss this trial with the patient.",
        "operationId": "approve_trial_referrals__referral_id__trials__trial_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "trial_id",
            "required": true,
            "schema": {
              "title": "Trial Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveTrialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Approver is not the referring physician"
          },
          "422": {
            "description": "Trial was not in the matched list for this case"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Approve Trial",
        "tags": [
          "trials"
        ]
      }
    },
    "/referrals/{referral_id}/waitlist": {
      "post": {
        "operationId": "waitlist_referrals__referral_id__waitlist_post",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Waitlist",
        "tags": [
          "referrals"
        ]
      }
    },
...
  }
}